Year 3 project in Computer Science Bachelor's at University of Paris Cité under the supervision of Professor Laurent Wendling
The code up to now contains a GUI and algorithms allowing the user to give an image, provide the location of a database (in nested directories format) and the program is able to find the closest matches using only the intersection of colour histograms. The effect of light is reduced by normalising the colours. Three main approaches are provided, classic colour histogram, grey histogram (slower due to the creation of a version of an histogram for every image in grey levels), and incremental intersection (fastest solution). This project provided proof of concept on the fact that the incremental version is a really fast and reliable option for the retrieval problem. This project can be used to achieve image segmentation.
This project is relevant in many fields, for example robot vision and understanding of their environment.